Bug fix.
Tested. Approved

Bug with the king and the trigger.
The bug occurred ONLY if we were using the application in the same exe
(i.e in the playing 2 or more instance in the editor). It would never
occur at the demo because each one would have its own exe and play only
its one...